Created a new folder called extensions to hold all the extension methods used. Created new files to hold the different kind of extensions. not a single file with thousands of lines.
